bugfix to enable MI 300 profiling
