+++ src/rocprof-compute
@@ -39,6 +39,8 @@ try:
     from rocprof_compute_base import RocProfCompute
     from utils.utils import console_error
 except ImportError as e:
+    import traceback
+    traceback.print_exc()
     # In wheel package, softlink is not supported.
     # rocprof-compute will get installed in bin and libexec/rocprofiler-compute.
     # When invoked from bin folder, the extra paths are required to import the dependent scripts
