+++ src/utils/mi_gpu_spec.py
@@ -4,7 +4,7 @@ from typing import Any, Dict
 
 import yaml
 
-from utils.logger import console_debug, console_error, console_warning
+from utils.logger import console_log, console_debug, console_warning, console_error
 
 # Constants for MI series
 # NOTE: Currently supports MI50, MI100, MI200, MI300
